# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Lightning Rod", "version": "0.1.0" },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HealthResponse" }
              }
            }
          }
        }
      }
    },
    "/lightning/latest-frame": {
      "get": {
        "summary": "Lightning Latest Frame",
        "operationId": "lightning_latest_frame_lightning_latest_frame_get",
        "parameters": [
          {
            "name": "satellite",
            "in": "query",
            "required": false,
            "schema": {
              "enum": ["goes-east", "goes-west"],
              "type": "string",
              "default": "goes-east",
              "title": "Satellite"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LightningFrameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/lightning/latest-points": {
      "get": {
        "summary": "Lightning Latest Points",
        "operationId": "lightning_latest_points_lightning_latest_points_get",
        "parameters": [
          {
            "name": "satellite",
            "in": "query",
            "required": false,
            "schema": {
              "enum": ["goes-east", "goes-west"],
              "type": "string",
              "default": "goes-east",
              "title": "Satellite"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 1000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LightningPointsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/imagery/cmi/ch13/frames": {
      "get": {
        "summary": "Get CMI Frames For A Time Window",
        "operationId": "cmi_ch13_frames_imagery_cmi_ch13_frames_get",
        "parameters": [
          {
            "name": "satellite",
            "in": "query",
            "required": false,
            "schema": {
              "enum": ["goes-east", "goes-west"],
              "type": "string",
              "description": "GOES satellite alias to query.",
              "default": "goes-east",
              "title": "Satellite"
            },
            "description": "GOES satellite alias to query."
          },
          {
            "name": "start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Inclusive UTC start timestamp for the requested playback window.",
              "title": "Start"
            },
            "description": "Inclusive UTC start timestamp for the requested playback window.",
            "example": "2026-04-08T00:00:00Z"
          },
          {
            "name": "end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Exclusive UTC end timestamp for the requested playback window.",
              "title": "End"
            },
            "description": "Exclusive UTC end timestamp for the requested playback window.",
            "example": "2026-04-08T01:00:00Z"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of frames to return within the requested time window.",
              "default": 1000,
              "title": "Limit"
            },
            "description": "Maximum number of frames to return within the requested time window."
          },
          {
            "name": "poll_hint",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 7200,
              "minimum": 1,
              "description": "Suggested client poll cadence in seconds for checking newer data availability.",
              "default": 3600,
              "title": "Poll Hint"
            },
            "description": "Suggested client poll cadence in seconds for checking newer data availability."
          }
        ],
        "responses": {
          "200": {
            "description": "CMI frame metadata for the requested time window, sorted oldest-to-newest.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CMIFramesResponse" }
              }
            }
          },
          "422": {
            "description": "Validation error. Common causes: missing `start`/`end`, invalid timestamps, or `start >= end`.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "description": "Returns rendered GOES ABI CMI Channel 13 frames for the requested UTC time window. This endpoint is time-window-only: clients must provide both `start` and `end`. Responses are ordered oldest-to-newest so clients can animate playback directly from the returned list."
      }
    },
    "/imagery/cmi/ch13/images/{satellite}/{frame_id}.png": {
      "get": {
        "summary": "Cmi Ch13 Image",
        "operationId": "cmi_ch13_image_imagery_cmi_ch13_images__satellite___frame_id__png_get",
        "parameters": [
          {
            "name": "satellite",
            "in": "path",
            "required": true,
            "schema": {
              "enum": ["goes-east", "goes-west"],
              "type": "string",
              "title": "Satellite"
            }
          },
          {
            "name": "frame_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Frame Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "PNG image for the requested CMI frame.",
            "content": { "image/png": {} }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CMIFrameModel": {
        "properties": {
          "frame_id": {
            "type": "string",
            "title": "Frame Id",
            "description": "Stable NOAA-derived frame identifier."
          },
          "satellite": {
            "type": "string",
            "title": "Satellite",
            "description": "Satellite alias requested by the client."
          },
          "start_time": {
            "type": "string",
            "title": "Start Time",
            "description": "Frame start timestamp in UTC ISO-8601 format."
          },
          "end_time": {
            "type": "string",
            "title": "End Time",
            "description": "Frame end timestamp in UTC ISO-8601 format."
          },
          "image_url": {
            "type": "string",
            "title": "Image Url",
            "description": "PNG URL for this rendered CMI frame."
          },
          "coordinates": {
            "items": {
              "prefixItems": [{ "type": "number" }, { "type": "number" }],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Coordinates",
            "description": "ImageSource corners ordered top-left, top-right, bottom-right, bottom-left."
          }
        },
        "type": "object",
        "required": [
          "frame_id",
          "satellite",
          "start_time",
          "end_time",
          "image_url",
          "coordinates"
        ],
        "title": "CMIFrameModel"
      },
      "CMIFramesResponse": {
        "properties": {
          "satellite": {
            "type": "string",
            "title": "Satellite",
            "description": "Satellite alias requested by the client."
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of frames returned for the requested time window."
          },
          "poll_interval_seconds": {
            "type": "integer",
            "title": "Poll Interval Seconds",
            "description": "Suggested cadence for checking whether newer frames may now be available."
          },
          "frames": {
            "items": { "$ref": "#/components/schemas/CMIFrameModel" },
            "type": "array",
            "title": "Frames",
            "description": "Frames sorted oldest-to-newest within the requested time window."
          }
        },
        "type": "object",
        "required": ["satellite", "count", "poll_interval_seconds", "frames"],
        "title": "CMIFramesResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": { "status": { "type": "string", "title": "Status" } },
        "type": "object",
        "required": ["status"],
        "title": "HealthResponse"
      },
      "LightningFeature": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "latitude": { "type": "number", "title": "Latitude" },
          "longitude": { "type": "number", "title": "Longitude" },
          "time": { "type": "string", "title": "Time" },
          "energy": {
            "anyOf": [{ "type": "number" }, { "type": "null" }],
            "title": "Energy"
          }
        },
        "type": "object",
        "required": ["id", "latitude", "longitude", "time"],
        "title": "LightningFeature"
      },
      "LightningFrameResponse": {
        "properties": {
          "frame_id": { "type": "string", "title": "Frame Id" },
          "satellite": { "type": "string", "title": "Satellite" },
          "start_time": { "type": "string", "title": "Start Time" },
          "end_time": { "type": "string", "title": "End Time" },
          "flash_count": { "type": "integer", "title": "Flash Count" },
          "updated_at": { "type": "string", "title": "Updated At" }
        },
        "type": "object",
        "required": [
          "frame_id",
          "satellite",
          "start_time",
          "end_time",
          "flash_count",
          "updated_at"
        ],
        "title": "LightningFrameResponse"
      },
      "LightningPointsResponse": {
        "properties": {
          "frame_id": { "type": "string", "title": "Frame Id" },
          "satellite": { "type": "string", "title": "Satellite" },
          "start_time": { "type": "string", "title": "Start Time" },
          "end_time": { "type": "string", "title": "End Time" },
          "updated_at": { "type": "string", "title": "Updated At" },
          "count": { "type": "integer", "title": "Count" },
          "features": {
            "items": { "$ref": "#/components/schemas/LightningFeature" },
            "type": "array",
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "frame_id",
          "satellite",
          "start_time",
          "end_time",
          "updated_at",
          "count",
          "features"
        ],
        "title": "LightningPointsResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}